fix sql: fix bad query and wrong type

+++ src/main/resources/db/migration/V10__create_table_transaction_payments.sql
@@ -6,7 +6,7 @@ CREATE TABLE transaction_payments (
     payment_to VARCHAR(100) NOT NULL,
     created_at TIMESTAMP NOT NULL,
     updated_at TIMESTAMP NOT NULL,
-    version INTEGER DEFAULT 0 NOT NULL,
+    version BIGINT DEFAULT 0 NOT NULL,
 
     CONSTRAINT fk_transaction_payments FOREIGN KEY (transaction_id)
         REFERENCES transactions(id)
